Move P2PDumpManager to statetransition util
